# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mapa de Emergencia y Rescate — API",
    "version": "1.0.0",
    "description": "API pública/admin del mapa de emergencia. Documentación generada automáticamente desde los bloques @swagger de cada route."
  },
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Lista de reportes de emergencia",
        "responses": {
          "200": {
            "description": "Reportes y bandera de persistencia"
          }
        }
      },
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Crear un reporte de emergencia",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "lat",
                  "lng",
                  "place",
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  },
                  "place": {
                    "type": "string"
                  },
                  "affected": {
                    "type": "integer"
                  },
                  "needs": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string",
                    "description": "data:image/...;base64 (opcional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reporte creado"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "No se pudo guardar"
          }
        }
      }
    },
...
    "/api/sync/geocode": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Geocodifica un lote de ubicaciones sin coordenadas (cron, requiere auth Bearer CRON_SECRET o token admin)",
        "parameters": [
          {
            "in": "query",
            "name": "max",
            "required": false,
            "description": "Tope de ubicaciones únicas a geocodificar en la corrida.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumen de la corrida de geocodificación.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al geocodificar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/reset": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Reinicia el cursor de sincronización (admin). No destructivo.",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "description": "ID de la fuente a reiniciar. Si se omite, reinicia todas.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor reiniciado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al reiniciar el cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}